Merge setSourceOpacity into set_text_visibility
In the original code, there was some duplication between these two functions and with the removal of custom fading, it did not make sense to keep both of them.

With the removal of setSourceOpacity, it left a bunch of other minor functions and variables unused, so these have been removed as well.